+++ src/models/mm_densenet.py
@@ -3,7 +3,7 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-from utils.utils_m_densenet import choose_layer_norm
+from utils.m_densenet import choose_layer_norm
 from models.transform import BandSplit
 from models.glu import GLU2d
 from models.m_densenet import MDenseNetBackbone, DenseBlock
